Removed paranoid code.
Configure Cesium Scene so that it throws errors during render to
catch bugs.

New scene.useWebXR flag to enable/disable WebXR code and revert to
WebVR emulated behaviour and make it easy to test both code paths
by avoiding to remake the CesiumJS library.

Use await instead of "cont" second-stage callbacks.

Better insight on the workings of Cesium.Scene allowed to remove
onXRFrameFirst which did processing that supposedly had to be done
until the first frame was executed.

Use const as much as possible, and if not, use let.